requirements.txt
python-dotenv~=1.0.1
requests~=2.32.3

langchain-core~=0.3.29
langchain~=0.3.13
langchain-community~=0.3.13
langchain-text-splitters~=0.3.4
langchain-openai~=0.2.14

gradio==5.9.1
gradio_client==1.5.2
